Changed default case for state table

+++ Lexer/lexergui/src/main/java/com/lexer/PrimaryController.java
@@ -111,7 +111,7 @@ public class PrimaryController {
 
         output.setItems(entries);
 
-        terminal.setText("Words Found: " + tokens.size() + "\nErrors Found: " + errorCount + "\nCorrect Rate: " + (tokens.size() - errorCount) / tokens.size() * 100 + "%");
+        terminal.setText("Words Found: " + tokens.size() + "\nErrors Found: " + errorCount + "\nCorrect Rate: " + (float)(tokens.size() - errorCount) / tokens.size() * 100.0 + "%");
     }
 
 }
